requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
colorama==0.4.6
coverage==6.5.0
Django==4.1.3
django-rest-framework==0.1.0
django-test==0.4030
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
exceptiongroup==1.0.4
flake8==5.0.4
iniconfig==1.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-watch==4.2.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
watchdog==2.1.9
